+++ server/src/main/kotlin/it/polito/server/degree/DegreeController.kt
@@ -32,8 +32,13 @@ class DegreeController (private val degreeService: DegreeService) {
     }
 
     @DeleteMapping("/{id}")
-    fun deleteDegree(@PathVariable id: String): ResponseEntity<Void> {
+    fun deleteDegree(@PathVariable id: String): ResponseEntity<Any> {
+        val degree = degreeService.findDegreeById(id)
+        if(degree == null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("ERROR, this Degree does NOT EXIST")
+
         degreeService.deleteDegree(id)
-        return ResponseEntity(HttpStatus.OK)
+        val successMessage = "Degree with ID $id successfully deleted."
+        return ResponseEntity.status(HttpStatus.OK).body(successMessage)
     }
 }
