+++ server/src/main/kotlin/it/polito/server/student/StudentController.kt
@@ -31,8 +31,13 @@ class StudentController(private val studentService: StudentService) {
     }
 
     @DeleteMapping("/{id}")
-    fun deleteStudent(@PathVariable id: String): ResponseEntity<Void> {
+    fun deleteStudent(@PathVariable id: String): ResponseEntity<Any> {
+        val student = studentService.findStudentById(id)
+        if(student==null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("ERROR, this Student does NOT EXIST")
         studentService.deleteStudent(id)
-        return ResponseEntity(HttpStatus.OK)
+        val successMessage = "Student with ID $id successfully deleted."
+        return ResponseEntity.status(HttpStatus.OK).body(successMessage)
+
     }
 }
